Bundler/Minifier

+++ src/GenerateServiceWorker.mjs
@@ -1,31 +1,30 @@
+import { existsSync } from "node:fs";
+import { FileFilter } from "./FileFilter.mjs";
 import { join } from "node:path/posix";
+import { SKIP_WAITING } from "../../flux-pwa/src/Pwa/SKIP_WAITING.mjs";
 import { readFile, writeFile } from "node:fs/promises";
-
-/** @typedef {import("./fileFilter.mjs").fileFilter} fileFilter */
-/** @typedef {import("../FluxPwaGenerator.mjs").FluxPwaGenerator} FluxPwaGenerator */
 
 export class GenerateServiceWorker {
     /**
-     * @type {FluxPwaGenerator}
+     * @type {FileFilter}
      */
-    #flux_pwa_generator;
+    #file_filter;
 
     /**
-     * @param {FluxPwaGenerator} flux_pwa_generator
      * @returns {GenerateServiceWorker}
      */
-    static new(flux_pwa_generator) {
+    static new() {
         return new this(
-            flux_pwa_generator
+            FileFilter.new()
         );
     }
 
     /**
-     * @param {FluxPwaGenerator} flux_pwa_generator
+     * @param {FileFilter} file_filter
      * @private
      */
-    constructor(flux_pwa_generator) {
-        this.#flux_pwa_generator = flux_pwa_generator;
+    constructor(file_filter) {
+        this.#file_filter = file_filter;
     }
 
     /**
@@ -34,61 +33,52 @@ export class GenerateServiceWorker {
      * @param {string} root
      * @param {string | null} application_cache_prefix
      * @param {{[key: string]: *} | null} data
-     * @param {fileFilter | null} file_filter
+     * @param {((root_file: string) => boolean) | null} file_filter
      * @param {boolean | null} exclude_jsdoc_files
      * @returns {Promise<void>}
      */
     async generateServiceWorker(service_worker_template_mjs_file, service_worker_mjs_file, root, application_cache_prefix = null, data = null, file_filter = null, exclude_jsdoc_files = null) {
         console.log(`Generate ${service_worker_mjs_file}`);
 
-        await writeFile(service_worker_mjs_file, "");
+        if (!existsSync(service_worker_mjs_file)) {
+            await writeFile(service_worker_mjs_file, "");
+        }
 
         const [
             files,
             excluded_file_filter_files,
             excluded_jsdoc_files
-        ] = await this.#flux_pwa_generator.scanFiles(
+        ] = await this.#file_filter.fileFilter(
             root,
             file_filter,
             exclude_jsdoc_files
         );
 
         if (excluded_file_filter_files.length > 0) {
-            console.log("- Excluded files from application cache files (File filter):");
-
             for (const root_file of excluded_file_filter_files) {
                 const file = join(root, root_file);
 
-                console.log(`  - ${file}`);
+                console.log(`Exclude ${file} from application cache files (File filter)`);
             }
         }
 
         if (excluded_jsdoc_files.length > 0) {
-            console.log("- Excluded files from application cache files (JSDoc file):");
-
             for (const root_file of excluded_jsdoc_files) {
                 const file = join(root, root_file);
 
-                console.log(`  - ${file}`);
+                console.log(`Exclude ${file} from application cache files (JSDoc file)`);
             }
         }
 
         await writeFile(service_worker_mjs_file, (await readFile(service_worker_template_mjs_file, "utf8")).replaceAll("{ /*%DATA%*/ }", JSON.stringify({
             ...data,
-            APPLICATION_CACHE_FILES: [
-                ...file_filter !== null && !file_filter(
-                    ""
-                ) ? [] : [
-                    ""
-                ],
-                ...files
-            ],
+            APPLICATION_CACHE_FILES: files,
             ...application_cache_prefix !== null ? {
                 APPLICATION_CACHE_PREFIX: application_cache_prefix
             } : null,
             APPLICATION_CACHE_VERSION: crypto.randomUUID(),
             ...application_cache_prefix !== null ? {
-                SKIP_WAITING: (await import("../../../flux-pwa/src/Pwa/SKIP_WAITING.mjs")).SKIP_WAITING
+                SKIP_WAITING
             } : null
         })));
     }
